+++ LICENSE
@@ -1,4 +1,4 @@
-The Answers React Components files listed in this repository are licensed under the below license.  All other features and products are subject to separate agreements
+The Search UI React files listed in this repository are licensed under the below license.  All other features and products are subject to separate agreements
 and certain functionality requires paid subscriptions to Yext products.
 
 Contains information from the language-subtag-registry JSON Database (https://github.com/mattcg/language-subtag-registry/tree/master/data/json)
