+++ src/utils/location-operations.ts
@@ -16,14 +16,14 @@ const defaultGeolocationOptions: PositionOptions = {
  * @public
  */
 export async function updateLocationIfNeeded(
-  SearchActions: SearchActions,
+  searchActions: SearchActions,
   intents: SearchIntent[],
   geolocationOptions?: PositionOptions
 ): Promise<void> {
-  if (intents.includes(SearchIntent.NearMe) && !SearchActions.state.location.userLocation) {
+  if (intents.includes(SearchIntent.NearMe) && !searchActions.state.location.userLocation) {
     try {
       const position = await getUserLocation(geolocationOptions);
-      SearchActions.setUserLocation({
+      searchActions.setUserLocation({
         latitude: position.coords.latitude,
         longitude: position.coords.longitude,
       });
