+++ tests/components/HierarchicalFacets.stories.tsx
@@ -1,6 +1,6 @@
 import { ComponentMeta } from '@storybook/react';
 import { SearchHeadlessContext, State } from '@yext/search-headless-react';
-import { generateMockedHeadless } from '../__fixtures__/answers-headless';
+import { generateMockedHeadless } from '../__fixtures__/search-headless';
 import { createHierarchicalFacet } from '../__utils__/hierarchicalfacets';
 import { RecursivePartial } from '../__utils__/mocks';
 import { HierarchicalFacets, HierarchicalFacetsProps } from '../../src';
